imprv colorshemes/catppuccin.sh

+++ colorschemes/catppuccin/catppuccin.sh
@@ -11,51 +11,51 @@ declare current_dir &&
 
 main() {
 
-    fisher install catppuccin/fish
-
-    fish_config theme save "Catppuccin Macchiato"
-
-    # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-
-    # Catppuccin for gnome-terminal
-    # see: https://github.com/catppuccin/gnome-terminal
-
-    curl -L https://raw.githubusercontent.com/catppuccin/gnome-terminal/v0.2.0/install.py | python3 -
+    if cmd_exists "fish"; then
+        fish -c "fisher install catppuccin/fish"
+        fish -c "fish_config theme save \"Catppuccin Macchiato\""
+    fi
 
     # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
     # Catppuccin for Bat
     # see: https://github.com/catppuccin/bat
 
-    git clone https://github.com/catppuccin/bat ~/.config/bat/catppuccin
+    if cmd_exists "bat"; then
+        git clone https://github.com/catppuccin/bat ~/.config/bat/catppuccin
 
-    cd ~/.config/bat/catppuccin || exit 1
+        cd ~/.config/bat/catppuccin || exit 1
 
-    mkdir -p "$(bat --config-dir)/themes"
+        mkdir -p "$(bat --config-dir)/themes"
 
-    cp *.tmTheme "$(bat --config-dir)/themes"
+        cp ./*.tmTheme "$(bat --config-dir)/themes"
 
-    bat cache --build
+        bat cache --build
+    fi
 
     # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
     # Catppuccin for lazygit
     # see: https://github.com/catppuccin/lazygit
 
-    git clone https://github.com/catppuccin/lazygit ~/.config/lazygit/catppuccin
+    if cmd_exists "lazygit"; then
+        git clone https://github.com/catppuccin/lazygit ~/.config/lazygit/catppuccin
 
-    cd ~/.config/lazygit/catppuccin || exit 1
+        cd ~/.config/lazygit/catppuccin || exit 1
 
-    local config_dir="$(lazygit --config-dir)"
+        config_dir="$(lazygit --config-dir)"
 
-    lazygit --use-config-file="${config_dir}/config.yml,macchiato-blue.yml"
+        lazygit --use-config-file="${config_dir}/config.yml,macchiato-blue.yml"
+    fi
 
     # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
     # Catppuccin for starship
     # see: https://github.com/catppuccin/starship
 
-    cp starship.toml ~/.config/starship.toml
+    if cmd_exists "starship"; then
+        cp starship.toml ~/.config/starship.toml
+    fi
 
 }
